requirements.txt
Django~=3.1.3
djangorestframework==3.12.2
psutil==5.7.3
django-ipware==3.0.1